Move resource binding checking to rrun
