+++ cpp/include/rrun/rrun.hpp
@@ -6,6 +6,8 @@
 #pragma once
 
 #include <optional>
+#include <string>
+#include <vector>
 
 namespace cucascade::memory {
 struct system_topology_info;
@@ -25,6 +27,94 @@ struct bind_options {
     bool network{true};  ///< Set `UCX_NET_DEVICES` to NICs near the GPU.
     bool verbose{false};  ///< Print warnings to stderr on binding failures.
 };
+
+/**
+ * @brief Live resource binding configuration collected from the running process.
+ *
+ * Holds the CPU affinity, NUMA memory binding, and network device configuration
+ * that are currently in effect. Obtained via `check_binding()`.
+ */
+struct resource_binding {
+    int rank = -1;  ///< Process rank (-1 if not available).
+    int gpu_id = -1;  ///< GPU device ID (-1 if not available).
+    std::string gpu_pci_bus_id;  ///< GPU PCI bus ID (empty if unavailable).
+    std::string cpu_affinity;  ///< CPU affinity string (e.g., "0-19,40-59").
+    std::vector<int> numa_nodes;  ///< NUMA node IDs bound to this process.
+    std::string ucx_net_devices;  ///< Value of the `UCX_NET_DEVICES` env var.
+};
+
+/**
+ * @brief Expected resource binding derived from topology information.
+ *
+ * Represents the binding configuration that *should* be in effect for a given
+ * GPU according to the system topology.
+ */
+struct expected_binding {
+    std::string cpu_affinity;  ///< Expected CPU affinity list.
+    std::vector<int> memory_binding;  ///< Expected NUMA node IDs.
+    std::vector<std::string> network_devices;  ///< Expected network devices.
+};
+
+/**
+ * @brief Results of validating actual vs. expected resource bindings.
+ */
+struct binding_validation {
+    bool cpu_ok = true;  ///< CPU affinity check passed.
+    bool numa_ok = true;  ///< NUMA binding check passed.
+    bool ucx_ok = true;  ///< UCX network devices check passed.
+    std::string expected_ucx_devices;  ///< Expected UCX devices (comma-separated).
+
+    /**
+     * @brief Check if all validations passed.
+     * @return true if CPU, NUMA, and UCX checks all passed.
+     */
+    [[nodiscard]] bool all_passed() const {
+        return cpu_ok && numa_ok && ucx_ok;
+    }
+};
+
+/**
+ * @brief Collect the live resource binding of the calling process.
+ *
+ * Queries the current CPU affinity, NUMA memory nodes, UCX network device
+ * configuration, process rank, and GPU information.
+ *
+ * @param gpu_id_hint  GPU device index hint.  When >= 0 the value is stored
+ *   directly; otherwise the GPU ID is read from `CUDA_VISIBLE_DEVICES`.
+ *   When a valid GPU ID is available, the PCI bus ID is also queried
+ *   (which may trigger CUDA initialization).
+ *
+ * @return The collected resource binding.
+ */
+resource_binding check_binding(int gpu_id_hint = -1);
+
+/**
+ * @brief Obtain the expected binding for a GPU from pre-discovered topology.
+ *
+ * Looks up @p gpu_id in @p topology and returns the expected CPU affinity,
+ * memory binding, and network devices.
+ *
+ * @param topology Pre-discovered system topology.
+ * @param gpu_id   GPU device index to look up.
+ * @return The expected binding, or `std::nullopt` if @p gpu_id is not found.
+ */
+std::optional<expected_binding> get_expected_binding(
+    cucascade::memory::system_topology_info const& topology, int gpu_id
+);
+
+/**
+ * @brief Validate an actual resource binding against an expected one.
+ *
+ * Compares the live @p actual binding with @p expected and reports per-resource
+ * pass/fail status.
+ *
+ * @param actual   Live resource binding (from `check_binding()`).
+ * @param expected Expected binding (from topology or a JSON file).
+ * @return Validation results.
+ */
+binding_validation validate_binding(
+    resource_binding const& actual, expected_binding const& expected
+);
 
 /**
  * @brief Bind the calling process to resources topologically close to a GPU.
